refactor routes in dedicated file
