+++ Makefile
@@ -1,7 +1,7 @@
 run: clean build
 	./bin/api
 build: clean
-	go build -o bin/api -v ./cmd/api/main.go
+	go build -o bin/api -v ./cmd/api/*.go
 clean:
 	rm -rf ./bin
 test:
